rename gob.py
